feat: refactor Pilkada 2024 page to use `Dashboard` component and update metadata TIL page file should be server component
