+++ modules/Pilkada2024/BoundaryLayers.tsx
@@ -4,7 +4,6 @@ import { useArea } from '@/hooks/useArea'
 import { config } from '@/lib/config'
 import { Area } from '@/lib/const'
 import dynamic from 'next/dynamic'
-import type { ReactElement } from 'react'
 import AreaBoundary from '../MapDashboard/AreaBoundary'
 import { useMapDashboard } from '../MapDashboard/hooks/useDashboard'
 import VotesChart from './VotesChart'
