+++ modules/Pilkada2024/VotesChart.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {
   type ChartConfig,
   ChartContainer,
