+++ modules/Pilkada2024/hooks/usePilkada.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useQuery } from '@tanstack/react-query'
 import {
   type PropsWithChildren,
